update: jcode 생성 삭제 시 transaction 및 비동기 처리
jcode 생성 및 삭제 시 trasaction을 걸어 master db가 처리하도록 설정
+ transaction 처리 시 시간이 오래 걸리므로 async 설정을 통해 비동기적으로 처리 설정
